requirements.txt
aiohttp
pylcs
python-dotenv
